--- a/engine/src/main/java/io/camunda/zeebe/engine/EngineConfiguration.java
+++ b/engine/src/main/java/io/camunda/zeebe/engine/EngineConfiguration.java
@@ -19,9 +19,13 @@ public final class EngineConfiguration {
   // This size (in bytes) is used as a buffer when filling an event/command up to the maximum
   // message size.
   public static final int BATCH_SIZE_CALCULATION_BUFFER = 1024 * 8;
+  public static final int DEFAULT_DMN_PARSED_DECISION_GRAPH_CACHE_CAPACITY = 5000;
 
   private int messagesTtlCheckerBatchLimit = DEFAULT_MESSAGES_TTL_CHECKER_BATCH_LIMIT;
   private Duration messagesTtlCheckerInterval = DEFAULT_MESSAGES_TTL_CHECKER_INTERVAL;
+
+  private int dmnParsedDecisionGraphCacheCapacity =
+      DEFAULT_DMN_PARSED_DECISION_GRAPH_CACHE_CAPACITY;
 
   public int getMessagesTtlCheckerBatchLimit() {
     return messagesTtlCheckerBatchLimit;
@@ -40,6 +44,16 @@ public final class EngineConfiguration {
   public EngineConfiguration setMessagesTtlCheckerInterval(
       final Duration messagesTtlCheckerInterval) {
     this.messagesTtlCheckerInterval = messagesTtlCheckerInterval;
+    return this;
+  }
+
+  public int getDmnParsedDecisionGraphCacheCapacity() {
+    return dmnParsedDecisionGraphCacheCapacity;
+  }
+
+  public EngineConfiguration setDmnParsedDecisionGraphCacheCapacity(
+      final int dmnParsedDecisionGraphCacheCapacity) {
+    this.dmnParsedDecisionGraphCacheCapacity = dmnParsedDecisionGraphCacheCapacity;
     return this;
   }
 }