## Description

**Before caching:**

Result "io.camunda.zeebe.engine.perf.EngineSingleDmnTaskPerformanceTest.measureProcessExecutionTime":
  3.749 ±(99.9%) 0.025 ms/op [Average]
  (min, avg, max) = (2.946, 3.749, 12.911), stdev = 0.761
  CI (99.9%): [3.724, 3.774] (assumes normal distribution)

Result "io.camunda.zeebe.engine.perf.EngineParallelDmnTaskPerformanceTest.measureProcessExecutionTime":
  26.837 ±(99.9%) 0.057 ms/op [Average]
  (min, avg, max) = (24.374, 26.837, 50.751), stdev = 1.744
  CI (99.9%): [26.779, 26.894] (assumes normal distribution)

**With parsed Dmn cache**

Result "io.camunda.zeebe.engine.perf.EngineSingleDmnTaskPerformanceTest.measureProcessExecutionTime":
  0.950 ±(99.9%) 0.016 ms/op [Average]
  (min, avg, max) = (0.651, 0.950, 11.944), stdev = 0.498
  CI (99.9%): [0.934, 0.967] (assumes normal distribution)

Result "io.camunda.zeebe.engine.perf.EngineParallelDmnTaskPerformanceTest.measureProcessExecutionTime":
  3.445 ±(99.9%) 0.023 ms/op [Average]
  (min, avg, max) = (2.746, 3.445, 23.453), stdev = 0.707
  CI (99.9%): [3.421, 3.468] (assumes normal distribution)
  
Summary:
**Single DMN Task Process P99.9 latency went down by 75%/ 4 times faster**
**10 Parallel DMN Task Process P99.9 latency went down by 87%/ 7-8 times faster**

## Related issues

relates #8571