@megglos 2023-08-03
Adopted to using a configurable LRU cache from agrona, performs slightly worse compared to the map.

Result "io.camunda.zeebe.engine.perf.EngineSingleDmnTaskPerformanceTest.measureProcessExecutionTime":
  1.020 ±(99.9%) 0.020 ms/op [Average]
  (min, avg, max) = (0.675, 1.020, 13.166), stdev = 0.617
  CI (99.9%): [1.000, 1.041] (assumes normal distribution)
  
 Result "io.camunda.zeebe.engine.perf.EngineParallelDmnTaskPerformanceTest.measureProcessExecutionTime":
  3.624 ±(99.9%) 0.032 ms/op [Average]
  (min, avg, max) = (2.777, 3.624, 21.808), stdev = 0.964
  CI (99.9%): [3.592, 3.656] (assumes normal distribution)