@megglos 2023-08-03
Guava Cache based, comparable to the agrona LRU:

Result "io.camunda.zeebe.engine.perf.EngineSingleDmnTaskPerformanceTest.measureProcessExecutionTime":
  **1.007** ±(99.9%) 0.017 ms/op [Average]
  (min, avg, max) = (0.653, 1.007, 13.365), stdev = 0.511
  CI (99.9%): [0.990, 1.023] (assumes normal distribution)

Result "io.camunda.zeebe.engine.perf.EngineParallelDmnTaskPerformanceTest.measureProcessExecutionTime":
  **3.651** ±(99.9%) 0.047 ms/op [Average]
  (min, avg, max) = (2.815, 3.651, 66.892), stdev = 1.436
  CI (99.9%): [3.604, 3.698] (assumes normal distribution)