@megglos 2023-08-03
Caffeine based, slightly better than agrona LRU and Guava:

Result "io.camunda.zeebe.engine.perf.EngineSingleDmnTaskPerformanceTest.measureProcessExecutionTime":
  0.997 ±(99.9%) 0.016 ms/op [Average]
  (min, avg, max) = (0.655, 0.997, 14.775), stdev = 0.483
  CI (99.9%): [0.981, 1.013] (assumes normal distribution)

Result "io.camunda.zeebe.engine.perf.EngineParallelDmnTaskPerformanceTest.measureProcessExecutionTime":
  3.586 ±(99.9%) 0.030 ms/op [Average]
  (min, avg, max) = (2.755, 3.586, 26.650), stdev = 0.913
  CI (99.9%): [3.556, 3.616] (assumes normal distribution)