@megglos 2023-08-03
As the agrona `Long2ObjectHashMap` map is not allowing us to set a limit I tested three ways to get LRU behavior.
Best performing seems Caffeine but it also requires us to introduce a new dependency, I may rather create a follow-up for considering replacing all caches by Caffeine for better performance.

Guava and agrona LRU where comparable but also the cache is not really under pressure with the current jmh test setup reusing the same dmn for all dmn Tasks, we may extend that to use different models.