@megglos 2023-08-03
Sync with @remcowesterhoud :
- the impl needs to change => cache within the DbDecisionState similar to like it's done for DbProcessState
- we aim to backport this to 8.2/8.1
- jmh tests will be omited for backporting
- ZPA takes this over to make it production ready
- the agrona LRU cache seems a good pick for now allowing us to already have the capacity limit