@remcowesterhoud 2023-08-07
> the agrona LRU cache seems a good pick for now allowing us to already have the capacity limit

For some reason the Agrona LRU doesn't seem to support removing entries from the cache manually. I'll go with the Guava option instead, as we already have the dependency in place and the performance was similar to Agrona.